Minor logging added.
The code is still quite complex to understand. Proper socket criteria needs to be established. Proper protocols needed to establish a network connection with TCP/IP.

Threading needs to be adjusted.

More info needs to be logged to understand how the program is running and to establish a proper debug protocol.